- Get ffs analysis reports
