--- src/modules/app/app.helpers.js
@@ -1,14 +1,0 @@
-const appHelpers = {
-  // groupFFSByHcpState(rows) {
-  //   return rows.reduce((acc, record) => {
-  //     if (!acc[record.hcp.state]) {
-  //       acc[record.hcp.state] = [record];
-  //     } else {
-  //       acc[record.hcp.state].push(record);
-  //     }
-  //     return acc;
-  //   }, {});
-  // },
-};
-
-export default appHelpers;
