+++ src/modules/app/app.helpers/index.js
@@ -1,0 +1,5 @@
+const { default: analysisFormatter } = require('./analysis.formatter');
+
+module.exports = {
+  analysisFormatter,
+};
